Migrate BBEH helper unit tests into behavior test; remove orphaned builtin/tests
assets/evaluators/builtin/tests/ was an orphaned unit suite using a sys.modules mocking hack. Its coverage is already provided by the per-asset behavior tests, except BBEH's low-level helpers (strip_latex/extract_answer/fuzzy_match), which are migrated into test_bbeh_evaluator_behavior.py (imported directly, no hack). ifeval (TestIFEvalInstructionCheckers/Util) and regex_match (comprehensive behavior test) unit tests were already redundant. Removes the folder and the tests.yml builtin_evaluators group (its only job ran the redundant regex_match unit test; e2e_test.py skips areas without tests.yml).
