--- assets/evaluators/builtin/tests/requirements.txt
@@ -1,4 +1,0 @@
-azure-ai-evaluation
-typing_extensions
-pytest
-pytest-cov
