+++ assets/evaluators/tests/test_evaluators_behavior/test_bbeh_evaluator_behavior.py
@@ -6,7 +6,110 @@
 import pytest
 from typing import Any, Dict
 from ..common.base_code_evaluator_runner import BaseCodeEvaluatorRunner
-from ...builtin.bbeh.evaluator._bbeh import BBEHEvaluator
+from ...builtin.bbeh.evaluator._bbeh import (
+    BBEHEvaluator,
+    strip_latex,
+    extract_answer,
+    fuzzy_match,
+)
+
+
+@pytest.mark.unittest
+class TestStripLatex:
+    """Unit tests for the strip_latex helper."""
+
+    def test_strip_dollar_signs(self):
+        """Test stripping dollar sign wrappers."""
+        assert strip_latex("$4$") == "4"
+        assert strip_latex("$answer$") == "answer"
+
+    def test_strip_boxed(self):
+        r"""Test stripping \boxed{} wrapper."""
+        assert strip_latex("\\boxed{4}") == "4"
+        assert strip_latex("\\boxed{answer}") == "answer"
+
+    def test_strip_text(self):
+        r"""Test stripping \text{} wrapper."""
+        assert strip_latex("\\text{answer}") == "answer"
+
+    def test_strip_texttt(self):
+        r"""Test stripping \texttt{} wrapper."""
+        assert strip_latex("\\texttt{code}") == "code"
+
+    def test_no_latex(self):
+        """Test that plain text is unchanged."""
+        assert strip_latex("plain text") == "plain text"
+        assert strip_latex("42") == "42"
+
+
+@pytest.mark.unittest
+class TestExtractAnswer:
+    """Unit tests for the extract_answer helper."""
+
+    def test_extract_the_answer_is(self):
+        """Test extracting with 'The answer is:' prefix."""
+        result = extract_answer("After analysis, The answer is: 42.")
+        assert result == "42"
+
+    def test_extract_the_final_answer_is(self):
+        """Test extracting with 'The final answer is' prefix."""
+        result = extract_answer("Let me think... The final answer is 25")
+        assert result == "25"
+
+    def test_extract_with_latex(self):
+        """Test extracting answer with LaTeX formatting."""
+        result = extract_answer("The final answer is: \\boxed{4}.")
+        assert result == "4"
+
+    def test_no_answer_prefix(self):
+        """Test that response without prefix returns trimmed version."""
+        result = extract_answer("Just 42.")
+        assert result == "Just 42"
+
+
+@pytest.mark.unittest
+class TestFuzzyMatch:
+    """Unit tests for the fuzzy_match helper."""
+
+    def test_exact_match(self):
+        """Test exact string match."""
+        assert fuzzy_match("answer", "answer") is True
+        assert fuzzy_match("42", "42") is True
+
+    def test_parenthesized_prediction(self):
+        """Test (a) matches a."""
+        assert fuzzy_match("(a)", "a") is True
+        assert fuzzy_match("(b)", "b") is True
+
+    def test_parenthesized_reference(self):
+        """Test a matches (a)."""
+        assert fuzzy_match("a", "(a)") is True
+        assert fuzzy_match("b", "(b)") is True
+
+    def test_numeric_equality(self):
+        """Test numeric comparison (4.0 == 4)."""
+        assert fuzzy_match("4.0", "4") is True
+        assert fuzzy_match("4", "4.0") is True
+        assert fuzzy_match("25.0", "25") is True
+
+    def test_quote_normalization(self):
+        """Test that single quotes are ignored."""
+        assert fuzzy_match("it's", "its") is True
+        assert fuzzy_match("don't", "dont") is True
+
+    def test_bracket_variations(self):
+        """Test [answer] matches answer."""
+        assert fuzzy_match("[42]", "42") is True
+        assert fuzzy_match("42", "[42]") is True
+
+    def test_question_mark_ending(self):
+        """Test trailing question mark is ignored."""
+        assert fuzzy_match("answer?", "answer") is True
+
+    def test_no_match(self):
+        """Test that different values don't match."""
+        assert fuzzy_match("yes", "no") is False
+        assert fuzzy_match("42", "43") is False
 
 
 @pytest.mark.unittest
